INital commit for Backend_User_Integration
