+++ app/src/main/java/appricottsoftware/clarity/RegisterActivity.java
@@ -9,6 +9,11 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
+import com.loopj.android.http.JsonHttpResponseHandler;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
+
 import java.io.UnsupportedEncodingException;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
@@ -16,14 +21,19 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 
+import appricottsoftware.clarity.sync.ClarityApp;
+import appricottsoftware.clarity.sync.ClarityClient;
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import cz.msebera.android.httpclient.Header;
 
 public class RegisterActivity extends AppCompatActivity implements View.OnClickListener {
 
     @BindView(R.id.bt_register) Button btRegister;
     @BindView(R.id.editText_email) EditText email;
     @BindView(R.id.editText_password) EditText password;
+
+    private static final String TAG = "RegisterActivity";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -42,6 +52,43 @@ public class RegisterActivity extends AppCompatActivity implements View.OnClickL
                 String hashedPassword = null;
                 hashedPassword = Hash_Password(password.getText().toString());
 
+                ClarityApp.getRestClient().RegisterRequest(email.getText().toString(), password.getText().toString(), this, new JsonHttpResponseHandler() {
+                    @Override
+                    public void onSuccess(int statusCode, Header[] headers, JSONObject response) {
+                        Log.e(TAG, "onSuccess1 : " + response.toString() );
+                        super.onSuccess(statusCode, headers, response);
+                    }
+
+                    @Override
+                    public void onSuccess(int statusCode, Header[] headers, JSONArray response) {
+                        Log.e(TAG, "onSuccess2 : " + response.toString());
+                        super.onSuccess(statusCode, headers, response);
+                    }
+
+                    @Override
+                    public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONObject errorResponse) {
+                        Log.e(TAG, "onFailue1 : " + errorResponse.toString());
+                        super.onFailure(statusCode, headers, throwable, errorResponse);
+                    }
+
+                    @Override
+                    public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONArray errorResponse) {
+                        Log.e(TAG, "onFailue2 : " + errorResponse.toString());
+                        super.onFailure(statusCode, headers, throwable, errorResponse);
+                    }
+
+                    @Override
+                    public void onFailure(int statusCode, Header[] headers, String responseString, Throwable throwable) {
+                        Log.e(TAG, "onFailue3 : " + responseString.toString());
+                        super.onFailure(statusCode, headers, responseString, throwable);
+                    }
+
+                    @Override
+                    public void onSuccess(int statusCode, Header[] headers, String responseString) {
+                        Log.e(TAG, "onSuccess3 : " + responseString.toString());
+                        super.onSuccess(statusCode, headers, responseString);
+                    }
+                });
 
                 Intent homeActivityIntent = new Intent(this, HomeActivity.class);
                 homeActivityIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK|Intent.FLAG_ACTIVITY_CLEAR_TOP);
