--- app/src/main/java/appricottsoftware/clarity/models/User.java
@@ -1,8 +1,0 @@
-package appricottsoftware.clarity.models;
-
-/**
- * Created by lepatrick714 on 1/29/18.
- */
-
-public class User {
-}
